This script will Categorise Reviews
This script will categorise the reviews through its extensive library of positive and negative words and then generate an insight.
It can be used with any database as long as it follows the metadata structure in the readme file attached.
Currently, the name of the app has no significance to the script as it only looks at the reviews column as it's made for one app at a time. (This just matched the structure of the dataset I used from Kaggle.)
Improvements that could be made to this is:
Faster runtime (maybe through multi-threading)
A larger collection of words can be associated with either a positive or negative review to give greater accuracy.
Removing the need for an app title at the start of every line OR incorporating it in so that it will give individual insights for individual apps.

P.S The dataset attached as a text file is the test set I used to check the accuracy of the script.